Cambiecito a timestamps

+++ Suse2/src/Suse2.c
@@ -119,7 +119,7 @@ void liberarDoblePuntero(char **puntero){
 uint64_t gettimeofday(){
 	struct timeval tv;
 	gettimeofday(&tv, NULL);
-	unsigned long long result = (((unsigned long long)tv.tv_sec)*1000 + ((unsigned long long)tv.tv_usec)/1000);
+	long long result = (((long long)tv.tv_sec)*1000 + ((long)tv.tv_usec)/1000);
 	uint64_t a = result;
 	return a;
 }
